requirements.txt
Pillow==2.2.1
argparse==1.2.1
numpy==1.8.0
wsgiref==0.1.2